+++ pkg/handler/license_handler_test.go
@@ -213,6 +213,7 @@ func TestLicenseHandler_GetComponentsLicense_ResponseStatus(t *testing.T) {
 		t.Fatalf("an error '%s' was not expected when opening a sugared logger", err)
 	}
 	config := &myconfig.ServerConfig{}
+	config.Lookup.SourcePriority = []int16{31, 32, 33, 5}
 	ctx := ctxzap.ToContext(context.Background(), zlog.L)
 	db, err := sqlx.Connect("sqlite", "file::memory:?cache=shared")
 	if err != nil {
@@ -303,6 +304,8 @@ func TestLicenseHandler_GetComponentLicense_ResponseStatus(t *testing.T) {
 		t.Fatalf("an error '%s' was not expected when opening a sugared logger", err)
 	}
 	config := &myconfig.ServerConfig{}
+	// Set look up priority
+	config.Lookup.SourcePriority = []int16{31, 32, 33, 5}
 	ctx := ctxzap.ToContext(context.Background(), zlog.L)
 	db, err := sqlx.Connect("sqlite", "file::memory:?cache=shared")
 	if err != nil {